Add Mastodon profile verification link
This entry in the head of the page enables link verification for the
link to this site in the Mastodon account profile. See:
https://docs.joinmastodon.org/user/profile/#verification

This will turn the "Website" link in https://fosstodon.org/@numba to
"verified" status with a green background and green tick, which provides
a little more authenticity to the account.

(Example of a profile with an already-verified website:
https://mastodon.social/@gmarkall)